refactorización del Setgame
