+++ src/containers/App/components/SetGame/setGame.jsx
@@ -2,19 +2,14 @@ import "./setGame.css";
 import React from "react";
 import Button from "../../../../components/Button/Button.jsx";
 import ConfigGame from "../../../../components/ConfigGame/ConfigGame.jsx";
-import { useSetGame } from "../../../../hooks/Setgame/useSetGame.js";
+import { useSetGame } from "../../../../services/Setgame/useSetGame.js";
 import { useNavigate } from "react-router-dom";
 
 //Botón para entrar a una partida + interfaz partida
 function Start() {
   const navigate = useNavigate();
-  const {
-    game_name,
-    setGameName,
-    max_players,
-    setMaxPlayers,
-    handleClick,
-  } = useSetGame(navigate);
+  const { game_name, setGameName, max_players, setMaxPlayers, handleClick } =
+    useSetGame(navigate);
 
   return (
     <div className="setGame__container">
@@ -24,14 +19,10 @@ function Start() {
           maxPlayers={max_players}
           setPlayerAmnt={setMaxPlayers}
         />
-        <Button
-          label="Crear Partida"
-          onClick={handleClick}
-        />
+        <Button label="Crear Partida" onClick={handleClick} />
       </div>
     </div>
   );
 }
-
 
 export default Start;
